add: comprehensive error handling for users and clothingitems controller

+++ controllers/clothingItems.js
@@ -1,32 +1,91 @@
+const {
+  BAD_REQUEST,
+  NOT_FOUND,
+  INTERNAL_SERVER_ERROR,
+  UNAUTHORIZED,
+} = require("../utils/errors");
 const ClothingItem = require("../models/clothingItem");
 
 // Get all clothing items
 const getAllClothes = (req, res) => {
+  if (!req.user?._id)
+    return res
+      .status(UNAUTHORIZED)
+      .send({ message: "Unauthorized to access resource!" });
+
   ClothingItem.find({})
-    .orFail()
-    .then((result) => {
-      res.send(result);
-    })
-    .catch((err) => res.send(err));
+    .then((result) => res.send(result))
+    .catch((err) => {
+      console.error(
+        `Error ${err.name} with the message ${err.message} has occurred while executing the code`
+      );
+      return res
+        .status(INTERNAL_SERVER_ERROR)
+        .send({ message: "An error has occurred on the server" });
+    });
 };
 
 // Create a new clothing item
 const createNewClothingItem = (req, res) => {
   const data = req.body;
-  data.owner = "67f82cbf3ab95914b0041f9a";
+
+  if (!req.user?._id)
+    return res
+      .status(UNAUTHORIZED)
+      .send({ message: "Unauthorized to create resource!" });
+
+  data.owner = req.user._id;
 
   ClothingItem.create(data)
-    .then((result) => res.send(result))
-    .catch((err) => res.send(err));
+    .then((result) => res.status(201).send(result))
+    .catch((err) => {
+      console.error(
+        `Error ${err.name} with the message ${err.message} has occurred while executing the code`
+      );
+      if (err.name === "ValidationError")
+        return res.status(BAD_REQUEST).send({ message: "Invalid data passed" });
+      else if (err.name === "CastError")
+        return res
+          .status(BAD_REQUEST)
+          .send({ message: "ID format is invalid" });
+      else
+        return res
+          .status(INTERNAL_SERVER_ERROR)
+          .send({ message: "An error has occurred on the server" });
+    });
 };
 
 // Delete a clothing item by id
 const deleteClothingItem = (req, res) => {
   const { itemId } = req.params;
 
+  if (!req.user?._id)
+    return res
+      .status(UNAUTHORIZED)
+      .send({ message: "Unauthorized to delete resource!" });
+
   ClothingItem.findByIdAndDelete(itemId)
-    .then((result) => res.send(result))
-    .catch((err) => res.send(err));
+    .orFail()
+    .then((result) => {
+      return res.send({ message: "Item deleted successfully!", item: result });
+    })
+    .catch((err) => {
+      console.error(
+        `Error ${err.name} with the message ${err.message} has occurred while executing the code`
+      );
+      if (err.name === "CastError")
+        return res
+          .status(BAD_REQUEST)
+          .send({ message: "ID format is invalid" });
+      else if (err.name === "DocumentNotFoundError")
+        return res
+          .status(NOT_FOUND)
+          .send({ message: `Clothing item with the id '${itemId}' not found` });
+      else
+        return res
+          .status(INTERNAL_SERVER_ERROR)
+          .send({ message: "An error has occurred on the server" });
+    });
 };
 
 module.exports = { getAllClothes, createNewClothingItem, deleteClothingItem };
